+++ CHANGES.rst
@@ -21,7 +21,8 @@ New features:
 
 Bug fixes:
 
-- *add item here*
+- Use zope.interface decorator.
+  [gforcada]
 
 
 4.1.2 (2016-06-07)